Library

MicroscopePSFs.Airy2D

MicroscopePSFs.Gauss2D

MicroscopePSFs.Gauss2D

MicroscopePSFs.InterpolatedPSF

MicroscopePSFs.PSF

MicroscopePSFs.PupilFunction

MicroscopePSFs.Scalar3D

MicroscopePSFs.SplinePSF

MicroscopePSFs.ZernikeCoefficients

MicroscopePSFs.importpsf

MicroscopePSFs.makeobs

MicroscopePSFs.nl2noll

MicroscopePSFs.nl2osa

MicroscopePSFs.noll2nl

MicroscopePSFs.noll2osa

MicroscopePSFs.normalize!

MicroscopePSFs.osa2nl

MicroscopePSFs.osa2noll

MicroscopePSFs.pdf

MicroscopePSFs.pdfₐ

MicroscopePSFs.radialpolynomial

MicroscopePSFs.zernikepolynomial

MicroscopePSFs.Airy2D — Type

Airy PSF

2D Airy Patter psf using paraxial, scalar model

#Fields

na : Numerical Aperture

λ : Emission Wavelength

pixelsize : Linear size of a back-projected pixel

The Airy PSF is I(r)=ν²/(4π)(2J₁(νr)/(νr))² where ν=πD/(λf)=2π*nₐ/λ

MicroscopePSFs.Gauss2D — Type

Gauss2D

Isotropic 2D Gaussian psf

#Fields

σ : Gaussian σ

pixelsize : Linear size of a pixel

MicroscopePSFs.Gauss2D — Method

Gauss2D(p::Airy2D)

convert an Airy2D to Gaussian PSF

MicroscopePSFs.PSF — Type

`PSF` defines an abstract type

MicroscopePSFs.PupilFunction — Type

PupilFunction

A structure used by 3D PSF methods.

#Fields

nₐ : Numerical Aperture

λ : Emission Wavelength

n : Immersion Index

pixelsize : Linear size of a back-projected pixel

'kpixelsize' : Number of pixels used in pupil function image

'pupil' : Pupil function ksizeksize2

Calculations of various 3D PSF types will involve calculations based around Pupil Funciton. They differ in the complexity and approach used to calculate Pupil Function.

MicroscopePSFs.Scalar3D — Type

Scalar3D

3D psf using scalar model and OTF rescaling

#Fields

pupilfunction : Pupil Function structure

'Σ' : OTF rescaling via image space 2D Gaussian Covariance matrix

'ksize' : number of pixels in pupil

MicroscopePSFs.SplinePSF — Type

SplinePSF(psfstack)

3D psf interpolated from other PSF models

#Fields

psfstack : a 3D psf stack

sitp : spline PSF object

MicroscopePSFs.importpsf — Method

importpsf(r)

load PSF data from PSF learning software

#Arguments

filename : file name of the PSF data, file format is .h5

#returns

p : Scalar3D PSF type

PSFstack : a 3D stack of learned PSF

pixelsize_x : pixel size at sample plane in x, micron

pixelsize_z : step size in z, micron

#Example: p, PSFstack, pixelsizex, pixelsizez = importpsf(filename)

MicroscopePSFs.makeobs — Method

makeobs(r)

make a set of observations points in a square windoe

#Arguments

r : A range

returns a vector of Tuples e.g. (x,y)

#Example: obs=makeobs(-8:8)

MicroscopePSFs.nl2noll — Method

nl2noll(n::Int,l::Int)

convert the `n` and `l` indexes into a Noll linear index

MicroscopePSFs.nl2osa — Method

nl2osa(n::Int,l::Int)

convert the `n` and `l` indexes into a OSE linear index

MicroscopePSFs.noll2nl — Method

noll2nl(j::Int)

convert the Noll index `j` into `n` and `l`

MicroscopePSFs.noll2osa — Method

noll2osa(j::Int)

convert the Noll index `j` to OSA index `j`

MicroscopePSFs.normalize! — Method

normalize!(p::PupilFunction)

normalize pupil using Parseval's theorem

MicroscopePSFs.osa2nl — Method

osa2nl(j::Int)

convert the OSA index `j` into `n` and `l`

MicroscopePSFs.osa2noll — Method

osa2noll(j::Int)

convert the OSA index `j` to Noll index `j`

MicroscopePSFs.pdf — Method

pdf(p::PSF,x_pixel::Tuple,x_emitter::Tuple)

return the psf at pixel x_pixel with emitter located at x_emitter.

#Arguments

p::PSF : psf structure

x_pixel::Tuple : location of pixel

x_emitter::Tuple : location of emitter

MicroscopePSFs.pdfₐ — Method

pdfₐ(p::PSF,x_pixel::Tuple,x_emitter::Tuple)

return the complex amplitude at pixel x_pixel with emitter located at x_emitter.

#Arguments

p::PSF : psf structure

x_pixel::Tuple : location of pixel

x_emitter::Tuple : location of emitter

MicroscopePSFs.radialpolynomial — Method

radialpolynomical(n::Int,m::Int,ρ)

return the value of the `n,m` radial polynomial at `ρ`

values of `ρ>1` will return zero

MicroscopePSFs.zernikepolynomial — Method

zernike(n::Int,l::Int,ρ,ϕ)

return the value the `n,l` zernike polynomial at `ρ,ϕ`

Note that l is in -n<:2:n